requirements.txt
pymupdf>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0.1
rapidfuzz>=3.6.0
